+++ datumbox-framework-core/src/main/java/com/datumbox/framework/core/machinelearning/common/abstracts/datatransformers/AbstractTransformer.java
@@ -59,9 +59,7 @@ public abstract class AbstractTransformer<MP extends AbstractTransformer.Abstrac
      */
     public void transform(Dataframe newData) {
         logger.info("transform()");
-        
-        knowledgeBase.init();
-        
+
         _convert(newData); 
         _normalize(newData);
     }
@@ -73,9 +71,7 @@ public abstract class AbstractTransformer<MP extends AbstractTransformer.Abstrac
      */
     public void denormalize(Dataframe data) {
         logger.info("denormalize()");
-        
-        knowledgeBase.init();
-        
+
         _denormalize(data);
     }