+++ datumbox-framework-core/src/main/java/com/datumbox/framework/core/machinelearning/common/abstracts/featureselectors/AbstractFeatureSelector.java
@@ -59,8 +59,6 @@ public abstract class AbstractFeatureSelector<MP extends AbstractFeatureSelector
     public void transform(Dataframe newData) {
         logger.info("transform()");
         
-        knowledgeBase.init();
-        
         _transform(newData);
     }